objectOracle/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Vision models
ultralytics==8.2.18        # YOLOv8
openai-clip==1.0.1
segment-anything==1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# Async & queuing
celery==5.4.0
redis==5.0.4
aiofiles==23.2.1

# Utils
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2
